general fixes

+++ include/project/montecarlo.hpp
@@ -8,8 +8,8 @@
 #include "hypersphere.hpp"
 #include "asset.hpp"
 
-std::pair<double, double> montecarloIntegration(int n, const std::string &function, HyperCube &hypercube, double &variance);
-std::pair<double, double> montecarloIntegration(int n, const std::string &function, HyperRectangle &hyperrectangle, double &variance);
-std::pair<double, double> montecarloIntegration(int n, const std::string &function, HyperSphere &hypersphere, double &variance);
+std::pair<double, double> montecarloIntegration(size_t n, const std::string &function, HyperCube &hypercube, double &variance);
+std::pair<double, double> montecarloIntegration(size_t n, const std::string &function, HyperRectangle &hyperrectangle, double &variance);
+std::pair<double, double> montecarloIntegration(size_t n, const std::string &function, HyperSphere &hypersphere, double &variance);
 
 #endif
